calculator클래스 테스트

+++ src/Java_221005/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
         return a-b;
     }
 
-    int multiple(int a, int b) {
+    double multiple(int a, int b) {
         return a*b;
     }
